+++ src/com/softserve/edu/less03/task1/AreaOfTriangle.java
@@ -1,4 +1,4 @@
-package com.softserve.edu.less2.task1;
+package com.softserve.edu.less03.task1;
 
 public class AreaOfTriangle {
     private double side1;
