+++ src/com/softserve/edu/less03/task2/Main2Task2.java
@@ -1,4 +1,4 @@
-package com.softserve.edu.less2.task2;
+package com.softserve.edu.less03.task2;
 
 import java.util.Scanner;
